+++ lib/drawGiftExchange.ts
@@ -1,5 +1,5 @@
-import { generateAndStoreSuggestions } from "@/app/api/giftSuggestions/route";
 import { SupabaseClient } from "@supabase/supabase-js";
+import { generateAndStoreSuggestions } from "./generateAndStoreSuggestion";
 
 // util function to draw gift exchange names
 // this function can be called from an API route
